fix(tests): extract FakeCaptureProbe for OpenCV integration tests
Made-with: Cursor
